+++ spec/lib/karafka/routing/topic_spec.rb
@@ -105,8 +105,9 @@ RSpec.describe_current do
     let(:expected_keys) do
       %i[
         kafka deserializer max_messages max_wait_time initial_offset id name active consumer
-        consumer_group_id subscription_group active_job consumer_persistence dead_letter_queue
-        declaratives manual_offset_management
+        consumer_group_id pause_max_timeout pause_timeout pause_with_exponential_backoff
+        subscription_group active_job consumer_persistence dead_letter_queue declaratives
+        manual_offset_management
       ]
     end